Release v0.15.0 BETA (#31)
* Bump Versions

* Update version links in changelog

* Just assume the port number is 5025 for now.

Fixes an issue where sometimes the port number is. Ideally we should change the type of `port` to be `number` instead of `string` and we should remove the `:` before the port number. I'm not sure what else depends on this `:` being there, though.
TSP-510

* Fix Issue when opening a new tsp-containing folder where `tspConfig` folder was not created

We weren't making sure the .vscode folder was created when trying to create the `tspConfig` folder under it.

TSP-549

* Be more explicit about the beta-ness of this extension

* Add known issue to connection not cleaning itself up

* Bump to kic-cli 0.15.0