Add back in the step that auto-publishes to the marketplace (#31)
In the v0.16.1 release, the version was automatically pushed before the
PR was closed (this was expected, it was part of testing). The step that
was added back in as part of this PR was removed to prevent naming
conflicts.

The publish step will only run after a successful merge from a release
branch.