TSP-839 Fix issue with naming of instrument connection (#64)
When we connect to an instrument, the information (io_type,
instr_address, model, serial_number, friendly_name etc.) is saved to
settings.json. Any item in Instruments pane (discovered or saved) is
initially created with label <model#serial_number>. For saved
instruments we need to update this to any connection_name if specified
by user and this information is fetched from settings.json. Since the
settings.json was not updated correctly due to not awaiting, the
Instruments pane showed wrong label as compared to terminal name. This
is fixed by awaiting when settings.json file is updated.

---------

Co-authored-by: GT, Shreya <shreya.gt@tektronix.com>