Add Setting to Dump Output Queue When Connecting (#71)
Sometimes developers may want TSP Toolkit to show the contents of the
output queue of the instrument when connecting instead of the queue
being cleared. This is especially helpful when connecting to an
instrument after the terminal is lost or the instrument drops the
connection.

This PR also improves the notifications shown while connecting to keep
the user informed of the progress of connecting to the instrument.